[Utilities] Move safe_plain_text_to_markdown to framework/utils.py and remove unused unescaping helper
