+++ internals/maintenance_scripts.py
@@ -18,7 +18,6 @@
 import collections
 import csv
 import logging
-import re
 from datetime import date, datetime, timedelta
 from io import StringIO
 from typing import Any
@@ -1796,25 +1795,3 @@ class DeleteWPTCoverageReport(FlaskHandler):
             f'Finished. Deleted a total of {count} old WPT coverage reports.'
         )
         return f'{count} WPT coverage reports deleted.'
-
-
-# ---------------------------------------------------------------------------
-# Idempotent Markdown Text Converters and Summary Migration Helpers
-# ---------------------------------------------------------------------------
-
-
-def safe_plain_text_to_markdown(text: str | None) -> str:
-    r"""Converts plain text to safe, idempotent markdown string.
-
-    Sanitizes HTML entities (< and >) to prevent XSS/ReDoS and prevents
-    double-escaping markdown symbols across multiple conversions.
-    """
-    if not text:
-        return ''
-
-    # 1. Sanitize HTML tags/entities (< -> &lt;, > -> &gt;)
-    safe_text = text.replace('<', '&lt;').replace('>', '&gt;')
-
-    # 2. Idempotently escape unescaped asterisks and underscores using negative lookbehind
-    safe_text = re.sub(r'(?<!\\)([\*_])', r'\\\1', safe_text)
-    return safe_text.strip()
